LayoutTests/imported/w3c/web-platform-tests/css/cssom/flex-serialization-expected.txt


FAIL CSSOM - Flex serialization assert_equals: Single value flex with CSS-wide keyword should serialize correctly. expected "flex: initial;" but got ""
